requirements.txt
-e .
Pillow==4.3.0
coverage==4.3.1

pytest==3.1.3
pytest-cov==2.5.1
pytest-django==3.1.2
flake8==3.4.1